+++ Homebanking/src/main/java/com/mindhub/Homebanking/models/Transaction.java
@@ -15,7 +15,7 @@ public class Transaction {
 
     private TransactionType type; // enum
 
-    private double amount;
+    private long amount;
     private String description;
     private LocalDateTime date;
     @ManyToOne(fetch = FetchType.EAGER)
@@ -26,7 +26,7 @@ public class Transaction {
     public Transaction() {
     }
 
-    public Transaction(TransactionType type, double amount, String description, LocalDateTime date) {
+    public Transaction(TransactionType type, long amount, String description, LocalDateTime date) {
         this.type = type;
         // Using the ternary operator, adjust the amount based on the type of transaction
         // If it is debit, the amount becomes negative; if it's credit, it stays positive
@@ -48,11 +48,11 @@ public class Transaction {
         this.type = type;
     }
 
-    public Double getAmount() {
+    public long getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(long amount) {
         this.amount = amount;
     }
